fix(tuval): give the browser entry its own import surface and tsconfig (#7836)
The page entry shared the kernel's import graph and reached `node:crypto` through
two barrels, so Vite's externalized `node:*` threw at module load and the desk
painted black. Split each shared slice into a `browser.ts` half beside its
`index.ts`, point the surface at it, and root the browser entry in
`tsconfig.browser.json` with `types: []` — so a `node:` import from the browser
surface is a plain `tsc` error, with no lint rule and no bundler plugin.
