+++ apps/tuval/src/shell/ui/Desk.tsx
@@ -19,8 +19,8 @@ import {useCallback, useEffect, useRef, useState} from "react";
 import type {ShellMsg, ShellState} from "../core/index.ts";
 import {activeWorkspace, processOf} from "../core/index.ts";
 import {defaultPrefixTable, type Key, type PrefixTable} from "../keys/index.ts";
-import type {PickerEntries} from "../picker/index.ts";
-import {noEntries} from "../picker/index.ts";
+import type {PickerEntries} from "../picker/browser.ts";
+import {noEntries} from "../picker/browser.ts";
 import {WindowId} from "../window/index.ts";
 import {CommandLine} from "./CommandLine.tsx";
 import {type ForwardedKey, ForwardedKeyProvider} from "./forwarded-key.tsx";
