+++ apps/tuval/tsconfig.browser.json
@@ -1,0 +1,17 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"noEmit": true,
+		"allowImportingTsExtensions": true,
+		// The browser entry's own lens. `types: []` is the whole guard this project exists for: with
+		// no `@types/node` in scope a `node:*` specifier resolves to nothing, so any module the entry
+		// reaches that imports one is a plain `tsc` error rather than a blank page (#7836).
+		"lib": ["ES2024", "DOM", "DOM.Iterable"],
+		"jsx": "react-jsx",
+		"types": [],
+		"tsBuildInfoFile": "./node_modules/.tmp/tsconfig.browser.tsbuildinfo"
+	},
+	// The entry alone. Everything else in the browser surface arrives by import, which is what makes
+	// this project's file set the entry's import graph and nothing wider.
+	"include": ["src/page/main.tsx", "src/page/assets.d.ts"]
+}
